ban and unban

+++ cogs/Moderation/ban.py
@@ -16,13 +16,23 @@ class ban(commands.Cog):
 
     @commands.has_permissions(ban_members=True)
     @commands.bot_has_permissions(ban_members=True)
-    @commands.command(description="Bans Member from the Server",aliases=['None'],usage=f"{os.path.basename(__file__)[:-3]} <user> []")
-    async def ban(self,ctx,user: discord.Member):
+    @commands.command(description="Bans Member from the Server",usage=f"{os.path.basename(__file__)[:-3]} <user> [reason]")
+    async def ban(self,ctx,user: discord.Member,*reason: str):
+        if(len(reason)!=0):
+            reason = " ".join([x for x in reason])
+        else:
+            reason=None
         view = confirmation.Buttons(ctx)
+        role = user.top_role
+        if(user==ctx.guild.owner):
+            await ctx.reply(f"{self.client.emotes['failed']} | Cannot Ban the Owner!")
+            return
+        await position_check(self,ctx,role)
         msg = await ctx.reply(f"You are about to ban: {user}",view=view)
         await view.wait()
         if view.value == "1":
             if msg: await msg.delete()
+            await ctx.guild.ban(user=user,reason=reason,delete_message_seconds=0)
             await ctx.reply(f"{self.client.emotes['success']} | {user} was banned successfully!")
             return False
         if view.value == "2":
@@ -31,4 +41,12 @@ class ban(commands.Cog):
             return False
 
 async def setup(client):
-    await client.add_cog(ban(client)) 
+    await client.add_cog(ban(client))   
+
+async def position_check(self, ctx, role):
+    if ctx.guild.get_member(self.client.user.id).top_role.position <= role.position:
+        await ctx.reply(f"{self.client.emotes['failed']} | My Role Isn't High Enough To Ban!")
+        raise commands.CommandError("Command Cancelled")
+    elif ctx.guild.owner_id != ctx.author.id and ctx.author.top_role.position <= role.position:
+        await ctx.reply(f"{self.client.emotes['failed']} | Your Role Isn't High Enough To Ban!")
+        raise commands.CommandError("Command Cancelled")
